Bump required Java version to Java 11
