+++ src/org/jetbrains/java/decompiler/util/ClasspathScanner.java
@@ -1,17 +1,24 @@
 // Copyright 2000-2017 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
 package org.jetbrains.java.decompiler.util;
 
+import java.lang.module.*;
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
+import java.nio.ByteBuffer;
 import java.io.File;
+import java.io.IOException;
+import java.io.InputStream;
 import java.util.HashSet;
+import java.util.Optional;
 import java.util.Set;
+import java.util.stream.Stream;
 
 import org.jetbrains.java.decompiler.main.DecompilerContext;
 import org.jetbrains.java.decompiler.main.extern.IFernflowerLogger.Severity;
 import org.jetbrains.java.decompiler.struct.StructContext;
 
 public class ClasspathScanner {
-    //TODO: Look into caching the metadata grabbed from the context info?
-    //J9 destroies the classpath so we need to scan this differently. We take advantage of MultiRelease jars to do this.
+
     public static void addAllClasspath(StructContext ctx) {
       Set<String> found = new HashSet<String>();
       String[] props = { System.getProperty("java.class.path"), System.getProperty("sun.boot.class.path") };
@@ -29,6 +36,45 @@ public class ClasspathScanner {
             ctx.addSpace(file, false);
             found.add(file.getAbsolutePath());
           }
+        }
+      }
+
+      addAllModulePath(ctx);
+    }
+
+    private static void addAllModulePath(StructContext ctx) {
+      for (ModuleReference module : ModuleFinder.ofSystem().findAll()) {
+        String name = module.descriptor().name();
+        try {
+          ModuleReader reader = module.open();
+          DecompilerContext.getLogger().writeMessage("Reading Module: " + name, Severity.INFO);
+          reader.list().forEach(cls -> {
+            if (!cls.endsWith(".class") || cls.contains("module-info.class"))
+              return;
+
+            DecompilerContext.getLogger().writeMessage("  " + cls, Severity.INFO);
+            try {
+              Optional<ByteBuffer> bb = reader.read(cls);
+              if (!bb.isPresent()) {
+                DecompilerContext.getLogger().writeMessage("    Error Reading Class: " + cls, Severity.ERROR);
+                return;
+              }
+
+              byte[] data;
+              if (bb.get().hasArray()) {
+                data = bb.get().array();
+              } else {
+                data = new byte[bb.get().remaining()];
+                bb.get().get(data);
+              }
+              ctx.addData(name, cls, data, false);
+            } catch (IOException e) {
+              DecompilerContext.getLogger().writeMessage("    Error Reading Class: " + cls, e);
+            }
+          });
+          reader.close();
+        } catch (IOException e) {
+          DecompilerContext.getLogger().writeMessage("Error loading module " + name, e);
         }
       }
     }
